// src/app/_dto/masters/billed-to/TblHBilledto.dto.ts
import { TbldBilledtoCo } from "./TbldBilledtoCo.dto";

export class TblHBilledto {
    HBilledto_SysID: number;
    HBilledto_Code: string;
    HBilledto_Name: string;
    HBilledto_Address_One: string;
    HBilledto_Address_Two: string;
    HBilledto_Address_Three: string;
    HBilledto_Address_Four: string;
    HBilledto_TRN_Number: string;
    HBilledto_Area_SysID: number;
    HBilledto_Area_Code: string;
    HBilledto_Area_Name: string;
    HBilledto_City_SysID: number;
    HBilledto_City_Code: string;
    HBilledto_City_Name: string;
    HBilledto_Count_SysID: number;
    HBilledto_Count_Code: string;
    HBilledto_Count_Name: string;
    HBilledto_Phone: string;
    HBilledto_Fax: string;
    HBilledto_Email_Address: string;
    HBilledto_AcDe_Yn: boolean;
    HBilledto_AcDe_SysID: number;
    HBilledto_AcDe_Code: string;
    HBilledto_AcDe_Name: string;

    TbldBilledtoCo:TbldBilledtoCo[] = []
  }
  